+++ src/lib/utils.ts
@@ -38,5 +38,5 @@ export const formatValue = (value: number, format: string) => {
   return new Intl.NumberFormat(undefined, {
     style: 'currency',
     currency: format,
-  }).format(parseInt(value.toFixed(2)))
+  }).format(parseFloat(value.toFixed(2)))
 }
